requirements.txt
attrs==24.2.0
blinker==1.6.2
Charm-Crypto @ file:///usr/local/src/charm-dev
click==8.1.7
exceptiongroup==1.1.3
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
greenlet==3.1.1
hypothesis==6.75.3
importlib_metadata==4.13.0
iniconfig==2.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
pyparsing==2.1.5
pytest==7.4.3
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.6.0
